split files up, fixed chapter model and fetch
